+++ source/extensions/lightspeed.trex.viewports.shared.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.8] - 2024-07-17
+### Changed
+- Updated test file path
+
 ## [1.2.7]
 ### Changed
 - Changed repo link